Consolidate getSearch and postSearch into single file
